requirements.txt
pandas
scikit-learn
umap-learn
hdbscan
# sentence-transformers 
bertopic # brew install llvm
langchain
langchain_ollama